Refactor Server Directory Architecture
