+++ src/main/java/com/stream/controller/UploadController.java
@@ -1,0 +1,26 @@
+package com.stream.controller;
+
+import com.stream.domain.video.dto.UploadVideoDto;
+import com.stream.facade.UploadFacade;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+@RestController
+public class UploadController {
+  @Autowired
+  private UploadFacade uploadFacade;
+
+  @PostMapping(path = "/videos/upload")
+  public ResponseEntity uploadVideo(
+          @Valid @RequestPart(value = "videoMetadata") UploadVideoDto videoMetadata,
+          @RequestPart(value = "videoFile") MultipartFile videoFile
+  ) {
+    uploadFacade.uploadVideoSync(videoMetadata, videoFile);
+    return ResponseEntity.ok().build();
+  }
+}
