+++ src/main/java/com/stream/storage/exception/FileStreamingException.java
@@ -1,0 +1,7 @@
+package com.stream.storage.exception;
+
+public class FileStreamingException extends RuntimeException {
+  public FileStreamingException(Exception cause) {
+    super("Occur Exception while create file stream to video streaming.", cause);
+  }
+}
