+++ src/main/java/com/stream/storage/exception/FileUploadException.java
@@ -1,0 +1,7 @@
+package com.stream.storage.exception;
+
+public class FileUploadException extends RuntimeException {
+  public FileUploadException(Exception cause) {
+    super("Occur Exception while file uploaded.", cause);
+  }
+}
